services/mongo_to_postgres/requirements.txt
pymongo==4.4.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
